README.md
Day 8 of 100 Days of Code. This is a Caesar Cipher which is a means to encrypt and decrypt messages. Typing encode will allow you to encode a message of your choice. Then, the user is asked for the shift amount before being presented with the encrypted message. Similarly, one can decode a message via the same method. The shift goes right if encoding and left if decoding. Special characters, white spaces, and numbers are ignored here.